+++ packages/react-components/src/index.ts
@@ -1,5 +1,6 @@
 export * from './useConnect';
 export * from './useConnection';
+export * from './useContractSchemaRpc';
 export * from './useContractSelector';
 export * from './useGrpcClient';
 export * from './useWalletConnectorSelector';